+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for [Allthemodium]
 
 ### Ported
+- Support for [Ex Nihilio Sequentia]
 - Support for [Industrial Foregoing]
 
 ## [6.0.5]