+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for [Allthemodium]
 
+### Changed
+- Updated support for [Alchemistry] to 2.2+
+
 ### Ported
 - Support for [Create]
 - Support for [Ex Nihilio Sequentia]
@@ -712,6 +715,7 @@ Added support for 18 more mods
 
 <!-- MODS -->
 [Allthemodium]: https://www.curseforge.com/minecraft/mc-mods/allthemodium
+[Alchemistry]: https://www.curseforge.com/minecraft/mc-mods/alchemistry
 [Applied Energistics]: https://www.curseforge.com/minecraft/mc-mods/applied-energistics-2
 [Ars Nouveau]: https://www.curseforge.com/minecraft/mc-mods/ars-nouveau
 [Astral Sorcery]: https://www.curseforge.com/minecraft/mc-mods/astral-sorcery