+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [6.1.0]
 ### Added
 - Support for [Allthemodium]
 - Support for [Elemental Craft]
@@ -645,7 +647,8 @@ Added support for 18 more mods
 - Added support for Extended Crafting
     - Combintation Core, Compressor, Ender Crafter and Tiered Tables
 
-[6.0.5]: https://github.com/TagnumElite/ProjectE-Integration/compare/v6.0.5...HEAD
+[6.1.0]: https://github.com/TagnumElite/ProjectE-Integration/compare/v6.1.0...HEAD
+[6.0.5]: https://github.com/TagnumElite/ProjectE-Integration/compare/v6.0.5...v6.1.0
 [6.0.4]: https://github.com/TagnumElite/ProjectE-Integration/compare/v6.0.4...v6.0.3
 [6.0.3]: https://github.com/TagnumElite/ProjectE-Integration/compare/v6.0.1...v6.0.3
 [6.0.1]: https://github.com/TagnumElite/ProjectE-Integration/compare/v5.3.2...v6.0.1